Epic/subjects

+++ backend/src/Gradebook.Foundation.Logic.Commands/FoundationCommands.cs
@@ -34,12 +34,12 @@ public class FoundationCommands : BaseLogic<IFoundationCommandsRepository>, IFou
 
         var invitation = invitationResult.Response!;
         if (invitation.IsUsed) return new StatusResponse<bool>(false, "Invitation already used");
-        if (invitation.ExprationDate < DateTime.Now) return new StatusResponse<bool>(false, "Invitation expired");
+        if (invitation.ExprationDate < DateTime.UtcNow) return new StatusResponse<bool>(false, "Invitation expired");
 
         var useInvitationResult = await Repository.UseInvitation(new UseInvitationCommand()
         {
             InvitationGuid = invitation.Guid,
-            UsedDate = DateTime.Now,
+            UsedDate = DateTime.UtcNow,
             UserGuid = userGuid.Response!
         });
         if (!useInvitationResult.Status) return new StatusResponse<bool>(false, useInvitationResult.Message);
@@ -73,7 +73,7 @@ public class FoundationCommands : BaseLogic<IFoundationCommandsRepository>, IFou
         var currentPerson = await _foundationQueries.Service.GetCurrentPersonGuid(command.SchoolGuid);
         if (!await _foundationPermissions.Service.CanCreateNewClass(currentPerson.Response))
             return new ResponseWithStatus<Guid>("Forbidden");
-        command.CreatedDate = DateTime.Now;
+        command.CreatedDate = DateTime.UtcNow;
         var resp = await Repository.AddNewClass(command);
         if (!resp.Status) return new ResponseWithStatus<Guid>(false, resp.Message);
         await Repository.SaveChangesAsync();
@@ -82,6 +82,9 @@ public class FoundationCommands : BaseLogic<IFoundationCommandsRepository>, IFou
 
     public async Task<ResponseWithStatus<Guid>> AddNewStudent(NewStudentCommand command, Guid schoolGuid)
     {
+        var person = await _foundationQueries.Service.GetCurrentPersonGuid(schoolGuid);
+        if (!person.Status) return new ResponseWithStatus<Guid>(false, "Cannot recognise person");
+        if (!await _foundationPermissions.Service.CanCreateNewStudents(person.Response)) return new ResponseWithStatus<Guid>("Person cannot create new students");
         Repository.BeginTransaction();
         command.CreatorGuid = (await _foundationQueries.Service.GetCurrentPersonGuid(schoolGuid)).Response;
         var resp = await Repository.AddNewStudent(command);
@@ -127,6 +130,7 @@ public class FoundationCommands : BaseLogic<IFoundationCommandsRepository>, IFou
         foreach (var student in studentsGuids)
         {
             var isStudentAlreadyInAnyClass = await _foundationQueries.Service.IsStudentInAnyClass(student);
+            if (!isStudentAlreadyInAnyClass.Status) return new StatusResponse("Cant check if students are able to assign to class");
             if (isStudentAlreadyInAnyClass.Response) return new StatusResponse(false, "At least one student is already assigned to class");
         }
         var currentPerson = await _foundationQueries.Service.RecogniseCurrentPersonByRelatedPerson(studentsGuids.First());
@@ -155,6 +159,10 @@ public class FoundationCommands : BaseLogic<IFoundationCommandsRepository>, IFou
 
     public async Task<StatusResponse> DeleteClass(Guid classGuid)
     {
+        var person = await _foundationQueries.Service.RecogniseCurrentPersonByClassGuid(classGuid);
+        if (!person.Status) return new StatusResponse(person.Message);
+        if (!await _foundationPermissions.Service.CanManageClass(classGuid, person.Response))
+            return new StatusResponse("Permission denied");
         var resp = await Repository.DeleteClass(classGuid);
         if (!resp.Status) return new StatusResponse(false, resp.Message);
         await Repository.SaveChangesAsync();
@@ -163,8 +171,21 @@ public class FoundationCommands : BaseLogic<IFoundationCommandsRepository>, IFou
 
     public async Task<StatusResponse> DeletePerson(Guid personGuid)
     {
+        var personToDelete = await _foundationQueries.Service.GetPersonByGuid(personGuid);
+        if (!personToDelete.Status) return new StatusResponse(personToDelete.Message);
+
+        var currentPerson = await _foundationQueries.Service.GetCurrentPersonGuid(personToDelete.Response!.SchoolGuid!.Value);
+        if (!currentPerson.Status) return new StatusResponse(currentPerson.Message);
+
+        if (personToDelete.Response!.SchoolRole.HasFlag(SchoolRoleEnum.Student))
+        {
+            if (!await _foundationPermissions.Service.CanDeleteStudents(currentPerson.Response))
+                return new StatusResponse("Permission denied");
+        }
+
         var resp = await Repository.DeletePerson(personGuid);
         if (!resp.Status) return new StatusResponse(false);
+
         await Repository.SaveChangesAsync();
         return new StatusResponse(true);
     }
@@ -334,5 +355,44 @@ public class FoundationCommands : BaseLogic<IFoundationCommandsRepository>, IFou
     {
         foreach (var guid in studentGuid) if (!(await Repository.RemoveStudentActiveClass(guid)).Status) return new StatusResponse(false, "Could not remove active school");
         return new StatusResponse(true);
+    }
+
+    public async Task<ResponseWithStatus<Guid>> AddSubject(Guid schoolGuid, NewSubjectCommand command)
+    {
+        var person = await _foundationQueries.Service.GetCurrentPersonGuid(schoolGuid);
+        if (!person.Status) return new ResponseWithStatus<Guid>(person.Message);
+        if (!await _foundationPermissions.Service.CanCreateNewSubject(person.Response))
+            return new ResponseWithStatus<Guid>("Permission denied");
+        var resp = await Repository.AddSubject(schoolGuid, command);
+        if (!resp.Status) return new ResponseWithStatus<Guid>(resp.Message);
+        await Repository.SaveChangesAsync();
+        return new ResponseWithStatus<Guid>(resp.Response);
+    }
+
+    public async Task<StatusResponse> EditTeachersInSubject(Guid subjectGuid, List<Guid> teachersGuids)
+    {
+        var currentPerson = await _foundationQueries.Service.GetCurrentPersonGuidBySubjectGuid(subjectGuid);
+        if (!currentPerson.Status) return new StatusResponse(currentPerson.Message);
+        if (!await _foundationPermissions.Service.CanManageSubject(subjectGuid, currentPerson.Response))
+            return new StatusResponse("Permission denied");
+        Repository.BeginTransaction();
+        var currentTeachersInSubject = await _foundationQueries.Service.GetTeachersForSubject(subjectGuid, 0);
+        if (!currentTeachersInSubject.Status) return new StatusResponse(currentTeachersInSubject.Message);
+        var currentTeachersInSubjectGuids = currentTeachersInSubject.Response!.Select(s => s.Guid);
+        var teachersToRemove = currentTeachersInSubjectGuids.Where(s => !teachersGuids.Contains(s));
+        var teachersToAdd = teachersGuids.Where(s => !currentTeachersInSubjectGuids.Contains(s));
+        var addResp = await Repository.AddTeachersToSubject(subjectGuid, teachersToAdd.ToList());
+        var removeResp = await Repository.RemoveTeachersFromSubject(subjectGuid, teachersToRemove.ToList());
+        if (addResp.Status && removeResp.Status)
+        {
+            await Repository.SaveChangesAsync();
+            Repository.CommitTransaction();
+            return new StatusResponse(true);
+        }
+        else
+        {
+            Repository.RollbackTransaction();
+            return new StatusResponse($"{addResp.Message}; {removeResp.Message}");
+        }
     }
 }
